+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### 功能构建
 
+- 优化数据源机制。
+  - 增加 com.dwarfeng.fdr.stack.service.SourceQosService。
+
 - 优化记录侧消费机制。
   - 优化 `[最大空闲时间] <= 0` 时的行为，使其更加符合直觉。